requirements.txt
# Dependencias principales
python-telegram-bot==20.7
python-dotenv==1.0.0

# Base de datos y utilidades
sqlite3

# Desarrollo y testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging y monitoreo
structlog==23.2.0

# Utilidades adicionales
python-dateutil==2.8.2
requests==2.31.0